test/owning-a-home/browser_testing/requirements.txt
selenium==2.45.0
nose==1.3.4
pyhamcrest==1.8.0
behave==1.2.5
requests==2.7.0
